--- a/hail/python/benchmark/table_benchmarks.py
+++ b/hail/python/benchmark/table_benchmarks.py
@@ -54,3 +54,16 @@ def table_big_aggregate_compile_and_execute():
     ht = hl.utils.range_table(M)
     expr = tuple([hl.agg.fraction(ht.idx % i == 0) for i in range(N) if i > 0])
     ht.aggregate(expr)
+
+def table_foreign_key_join(n1: int, n2: int):
+    ht = hl.utils.range_table(n1)
+    ht2 = hl.utils.range_table(n2)
+    ht.annotate(x = ht2[(n1 - 1 - ht.idx) % n2])._force_count()
+
+@benchmark
+def table_foreign_key_join_same_cardinality():
+    table_foreign_key_join(1_000_000, 1_000_000)
+
+@benchmark
+def table_foreign_key_join_left_higher_cardinality():
+    table_foreign_key_join(1_000_000, 1_000)