--- a/hail/python/hail/matrixtable.py
+++ b/hail/python/hail/matrixtable.py
@@ -2387,7 +2387,7 @@ class MatrixTable(ExprContainer):
             if is_row_key:
                 def joiner(left):
                     return MatrixTable(MatrixAnnotateRowsTable(
-                        left._mir, right.rows()._tir, uid, None))
+                        left._mir, right.rows()._tir, uid))
                 schema = tstruct(**{f: t for f, t in self.row.dtype.items() if f not in self.row_key})
                 ir = Join(GetField(TopLevelReference('va'), uid),
                           uids_to_delete,