--- a/hail/src/test/scala/is/hail/variant/vsm/PartitioningSuite.scala
+++ b/hail/src/test/scala/is/hail/variant/vsm/PartitioningSuite.scala
@@ -36,8 +36,7 @@ class PartitioningSuite extends SparkSuite {
       MatrixAnnotateRowsTable(
         ir.MatrixRead(rangeReader.fullType, false, false, rangeReader),
         t,
-        "foo",
-        None))
+        "foo"))
       .rvd.count()
   }