Some context:

We have 3 kinds of IRs currently: 
 - Value IRs (MakeStruct, I32, ApplyComparisonOp, etc)
 - TableIRs (TableRange, TableFilter, etc)
 - MatrixIRs (MatrixMapRows, MatrixExplodeCols, etc)

One of the passes of the optimizer is to reformulate MatrixIRs in terms of TableIRs, and we're in the middle of this push to write lowerers for each MatrixIR node (which means writing an algorithm in LowerMatrixTable.scala, and removing the node's 'execute' method).

Here I implement lowering MatrixAnnotateRowsTable as either TableIntervalJoin or TableLeftJoinRightDistinct (a future PR should split MatrixAnnotateRowsTable into 2 nodes like Table has, probably).

The one bit of extra complexity in Python comes from implementing foreign-key joins explicitly (this was previously handled by the IR node itself).