@akotlar 2019-01-06
Thanks for the context! Unit tests pass, nice. A few warnings generated: `assertEquals` is deprecated in favor of `assertEqual` (I'm using pytest-4.0.2/python 3.7.1 may be new deprecation)

Need to spend more time following the move of key from IR to table. Do the existing join unit tests cover the new join implementation?