@akotlar 2019-01-11
I think this is good to go, once MatrixIR.scala comments pertaining to execute are removed. I would like to know, as an aside, more about execute.

Coverage of modified join functionality seems good!

Breaking line 1505, using
```python
 join_table = src.rows()
```

generates a test error in
test/hail/matrixtable/test_matrix_table.py:490

Breaking line 1529, using
```python
joiner = lambda left: 1
```
triggers an error in test/hail/matrixtable/test_matrix_table.py:905, which tests single row