@tpoterba 2019-01-14
we have a benchmarking suite in Python - it's very early, though, and not run as part of the CI. I'll add something there.