@akotlar 2019-01-14
Thank you both. If you're comfortable without performance testing at this stage (since this is part of a broader move, and I optimization seems like it is out of scope, as it is mostly now a function of the Table implementation), this PR is approved, once the commented-out old code block is removed.