@tpoterba 2019-01-14
Code:

```
In [2]: def table_foreign_key_join(n1: int, n2: int):
   ...:     ht = hl.utils.range_table(n1)
   ...:     ht2 = hl.utils.range_table(n2)
   ...:     ht.annotate(x = ht2[(n1 - 1 - ht.idx) % n2])._force_count()
   ...:

In [3]: %timeit table_foreign_key_join(1000000, 1000000)
```

Timings:

```
Master:
12.6 s ± 191 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

PR:
14.1 s ± 266 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```