@tpoterba 2019-01-14
It is horribly depressing that this takes more than 10 seconds.